feature: adjust detekt and ktlint
